Move RefSpaceFEFieldFunction to separate file and NonMatching namespace
